serde: Don't pull in default features
They are not used, so let's not build them.

(Also, some crates still cause trouble when serde is active with std
when the expect it not to be, so let's not cause any of these troubles
needlessly).

See-Also: https://gitlab.com/chrysn/coap-handler/-/issues/1
See-Also: https://github.com/KokaKiwi/rust-hex/pull/42